IsGrounded is fixed!!!
Now I need to do a double jump and destroy the bullet prefabs and make them travel.

It took me a while to figure out the boxCast check but eventually it was solved because my o for the OnDrawGizmos function was not capital so it was not visible. Coding hurts. But fixed that an ran into an issue for the IsGrounded not working anymore only to find out I removed the brackets so the OR function was overriding the AND function in the if statement. Anyways, we are moving on now!